move AflTargetArgs to libafl_bolts and adapt CMainArgsBuilder to use it
