+++ libafl_bolts/src/lib.rs
@@ -113,9 +113,14 @@ pub mod subrange;
 pub mod tuples;
 
 #[cfg(all(feature = "std", unix))]
-pub mod cargs;
+pub mod argparse;
 #[cfg(all(feature = "std", unix))]
-pub use cargs::*;
+pub use argparse::*;
+
+#[cfg(all(feature = "std", unix))]
+pub mod afl_args;
+#[cfg(all(feature = "std", unix))]
+pub use afl_args::*;
 
 /// The purpose of this module is to alleviate imports of the bolts by adding a glob import.
 #[cfg(feature = "prelude")]
